// src/api-client/api.ts
export interface LocationWeatherDto {
    /**
     * 
     * @type {string}
     * @memberof LocationWeatherDto
     */
    'location': string;
    /**
     * 
     * @type {string}
     * @memberof LocationWeatherDto
     */
    'countryCode': string;
    /**
     * 
     * @type {Array<WeatherDto>}
     * @memberof LocationWeatherDto
     */
    'weather': Array<WeatherDto>;
}
/**
 * 
 * @export
 * @interface WeatherDto
 */
export interface WeatherDto {
    /**
     * 
     * @type {string}
     * @memberof WeatherDto
     */
    'timestamp': string;
    /**
     * 
     * @type {number}
     * @memberof WeatherDto
     */
    'temperatureCelsius': number;
    /**
     * 
     * @type {number}
     * @memberof WeatherDto
     */
    'pressure': number;
    /**
     * 
     * @type {number}
     * @memberof WeatherDto
     */
    'humidity': number;
    /**
     * 
     * @type {number}
     * @memberof WeatherDto
     */
    'windSpeedMetersPerSecond': number;
    /**
     * 
     * @type {number}
     * @memberof WeatherDto
     */
    'windDirection': number;
    /**
     * 
     * @type {number}
     * @memberof WeatherDto
     */
    'rainVolumePast3HoursMm'?: number;
    /**
     * 
     * @type {number}
     * @memberof WeatherDto
     */
    'snowVolumePast3HoursMm'?: number;
    /**
     * 
     * @type {number}
     * @memberof WeatherDto
     */
    'precipitationProbability': number;
    /**
     * 
     * @type {string}
     * @memberof WeatherDto
     */
    'weatherDescription': string;
    /**
     * 
     * @type {boolean}
     * @memberof WeatherDto
     */
    'isNight': boolean;
}